requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
cryptography==2.6.1
dj-database-url==0.5.0
Django==2.2.1
django-cors-headers==3.0.2
django-environ==0.4.5
django-heroku==0.3.1
django-rest-knox==4.0.1
djangorestframework==3.9.3
gunicorn==19.9.0
psycopg2==2.8.2
pycparser==2.19
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1.2
django-environ==0.4.5
pandas==0.24.2
